+++ src/main/java/ru/wilyfox/client/protocol/DwBossDamagePacket.java
@@ -2,6 +2,6 @@ package ru.wilyfox.client.protocol;
 
 public record DwBossDamagePacket(
         String bossId,
-        long damage
+        int damage
 ) {
 }
